Add files via upload
This repository contains programs that show how to work with files in C++. It covers how to read from and write to files, as well as how to create, open, and close files. File handling helps you manage data outside your program and is useful for saving and loading information.